+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod client;
 pub mod codec;
+pub mod metadata;
 pub mod server;
 pub mod status;
 pub mod transport;